Deny GFXOFF in amdgpu_ring_commit using PG state change
When compute and gfx rings are active at the same time under
some conditions the GPU hang occurs. To prevent this we need
to perform powergating state change from GATE to UNGATE before
ringing doorbell and restore it back to GATE afterwards.

Make sure the GFXOFF is denied when ringing doorbell for GFX IPs:
GFX and SDMA. For COMPUTE IP powergate state change is required.